Fix dropdown component in grid
